requirements.txt
Django >= 1.5
MySQL-python
South
django-tastypie
